Predictive Train Delay and Disruption Analysis System/backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.21.1
requests==2.31.0

pydantic==1.10.9
python-dotenv==1.0.0

numpy==1.26.4
pandas==2.1.2
scikit-learn==1.3.2
xgboost==1.7.6
shap==0.42.1

sqlalchemy==2.0.19
psycopg2-binary==2.9.7
fastapi
uvicorn[standard]
requests
python-multipart
pydantic
aiohttp
python-dotenv